lab1 refactor

+++ lab1/src/Car/Car.java
@@ -1,5 +1,6 @@
 package Car;
 
+import java.math.BigDecimal;
 import java.util.Scanner;
 
 public class Car {
@@ -7,10 +8,10 @@ public class Car {
     private int id;
     private String model;
     private int year;
-    private double price;
+    private BigDecimal price;
     private String regNumber;
 
-    public Car(int id, String model, int year, double price, String regNumber){
+    public Car(int id, String model, int year, BigDecimal price, String regNumber){
         this.id = id;
         this.model = model;
         this.year = year;
@@ -31,7 +32,7 @@ public class Car {
         year = scanner.nextInt();
 
         System.out.println("Enter car price:");
-        price = scanner.nextDouble();
+        price = scanner.nextBigDecimal();
 
         System.out.println("Enter car register number:");
         regNumber = scanner.next();
@@ -40,7 +41,6 @@ public class Car {
     public int getId() {
         return id;
     }
-
     public void setId(int id) {
         this.id = id;
     }
@@ -48,7 +48,6 @@ public class Car {
     public String getModel() {
         return model;
     }
-
     public void setModel(String model) {
         this.model = model;
     }
@@ -56,23 +55,20 @@ public class Car {
     public int getYear() {
         return year;
     }
-
     public void setYear(int year) {
         this.year = year;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
-
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
 
     public String getRegNumber() {
         return regNumber;
     }
-
     public void setRegNumber(String regNumber) {
         this.regNumber = regNumber;
     }
